feat(feature:about): Migrate about module to CMP
