+++ feature/about/src/androidMain/kotlin/org/mifos/mobile/feature/about/OpenUrl.android.kt
@@ -1,0 +1,17 @@
+/*
+ * Copyright 2025 Mifos Initiative
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * See https://github.com/openMF/mobile-mobile/blob/master/LICENSE.md
+ */
+package org.mifos.mobile.feature.about
+
+import android.content.Intent
+import android.net.Uri
+
+actual fun openUrl(url: String?) {
+    val intent = Intent(Intent.ACTION_VIEW, Uri.parse(url))
+}
