+++ feature/about/src/commonMain/composeResources/drawable/feature_about_law_icon.xml
