+++ feature/about/src/commonMain/composeResources/drawable/feature_about_source_code.xml
