+++ feature/about/src/commonMain/composeResources/drawable/mifos_logo.png
