+++ feature/about/src/desktopMain/kotlin/org/mifos/mobile/feature/about/OpenUrl.desktop.kt
@@ -1,0 +1,19 @@
+/*
+ * Copyright 2025 Mifos Initiative
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * See https://github.com/openMF/mobile-mobile/blob/master/LICENSE.md
+ */
+package org.mifos.mobile.feature.about
+
+import java.awt.Desktop
+import java.net.URI
+
+actual fun openUrl(url: String?) {
+    if (Desktop.isDesktopSupported()) {
+        Desktop.getDesktop().browse(URI(url))
+    }
+}
