+++ feature/about/src/jsMain/kotlin/org/mifos/mobile/feature/about/OpenUrl.js.kt
@@ -1,0 +1,14 @@
+/*
+ * Copyright 2025 Mifos Initiative
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * See https://github.com/openMF/mobile-mobile/blob/master/LICENSE.md
+ */
+package org.mifos.mobile.feature.about
+
+actual fun openUrl(url: String?) {
+//    window.open(url, "_blank") // Opens in a new tab"
+}
