--- feature/about/src/main/java/org/mifos/mobile/feature/about/ui/AboutUsHeader.kt
@@ -1,64 +1,0 @@
-/*
- * Copyright 2024 Mifos Initiative
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * See https://github.com/openMF/mobile-mobile/blob/master/LICENSE.md
- */
-package org.mifos.mobile.feature.about.ui
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import org.mifos.mobile.core.ui.utils.DevicePreviews
-import org.mifos.mobile.feature.about.R
-
-@DevicePreviews
-@Composable
-internal fun AboutUsHeader(
-    modifier: Modifier = Modifier,
-) {
-    Column(modifier) {
-        Image(
-            painter = painterResource(id = R.drawable.mifos_logo),
-            contentDescription = null,
-            modifier = Modifier
-                .size(100.dp)
-                .fillMaxWidth()
-                .align(Alignment.CenterHorizontally),
-        )
-
-        Text(
-            text = stringResource(id = R.string.feature_about_app_name),
-            style = MaterialTheme.typography.headlineMedium,
-            color = MaterialTheme.colorScheme.onSurface,
-            textAlign = TextAlign.Center,
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(bottom = 8.dp),
-        )
-
-        Text(
-            text = stringResource(id = R.string.feature_about_description),
-            style = MaterialTheme.typography.titleSmall.copy(),
-            color = MaterialTheme.colorScheme.onSurface,
-            textAlign = TextAlign.Center,
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(bottom = 16.dp),
-        )
-    }
-}
