--- feature/about/src/main/java/org/mifos/mobile/feature/about/ui/PrivacyPolicyScreen.kt
@@ -1,106 +1,0 @@
-/*
- * Copyright 2024 Mifos Initiative
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * See https://github.com/openMF/mobile-mobile/blob/master/LICENSE.md
- */
-package org.mifos.mobile.feature.about.ui
-
-import android.annotation.SuppressLint
-import android.content.Context
-import android.content.Intent
-import android.graphics.Bitmap
-import android.net.Uri
-import android.webkit.WebView
-import android.webkit.WebViewClient
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.height
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.viewinterop.AndroidView
-import org.mifos.mobile.core.designsystem.components.MifosScaffold
-import org.mifos.mobile.core.ui.component.MifosProgressIndicator
-import org.mifos.mobile.feature.about.R
-
-@Composable
-internal fun PrivacyPolicyScreen(
-    navigateBack: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val context = LocalContext.current
-    MifosScaffold(
-        topBarTitleResId = R.string.feature_about_privacy_policy,
-        navigateBack = navigateBack,
-        modifier = modifier,
-        content = {
-            WebView(
-                url = context.getString(R.string.feature_about_policy_url),
-            )
-        },
-    )
-}
-
-@SuppressLint("SetJavaScriptEnabled")
-@Composable
-private fun WebView(
-    url: String,
-    modifier: Modifier = Modifier,
-) {
-    var isLoading by remember { mutableStateOf(true) }
-
-    Column(modifier) {
-        Spacer(modifier = Modifier.height(20.dp))
-        AndroidView(
-            factory = { context ->
-                WebView(context).apply {
-                    settings.javaScriptEnabled = true
-                    settings.domStorageEnabled = true
-                    overScrollMode = WebView.OVER_SCROLL_NEVER
-                    this.webViewClient = object : WebViewClient() {
-                        @Deprecated("Deprecated in Java")
-                        override fun shouldOverrideUrlLoading(
-                            view: WebView,
-                            url: String,
-                        ): Boolean {
-                            return if (context.checkUri(url)) {
-                                false
-                            } else {
-                                context.startActivity(Intent(Intent.ACTION_VIEW, Uri.parse(url)))
-                                true
-                            }
-                        }
-
-                        override fun onPageStarted(view: WebView?, url: String?, favicon: Bitmap?) {
-                            isLoading = true
-                        }
-
-                        override fun onPageFinished(view: WebView?, url: String?) {
-                            isLoading = false
-                        }
-                    }
-                    loadUrl(url)
-                }
-            },
-            update = { webView ->
-                webView.loadUrl(url)
-            },
-        )
-        if (isLoading) {
-            MifosProgressIndicator()
-        }
-    }
-}
-
-private fun Context.checkUri(uri: String): Boolean {
-    return Uri.parse(uri)?.host?.endsWith(this.getString(R.string.feature_about_policy_host)) == true
-}
